Separate TypeHelpers.h and constant evaluation unit tests for bit-stream casting.
