+++ source/symbols/Type.cpp
@@ -6,7 +6,7 @@
 //------------------------------------------------------------------------------
 #include "slang/symbols/Type.h"
 
-#include <numeric>
+#include "TypeHelpers.h"
 
 #include "slang/compilation/Compilation.h"
 #include "slang/diagnostics/LookupDiags.h"
@@ -480,59 +480,6 @@ bool Type::isAssignmentCompatible(const Type& rhs) const {
     return false;
 }
 
-static std::pair<bitwidth_t, bitwidth_t> linearCoefficients(const Type& type,
-                                                            unsigned int destination = 0) {
-    /* IEEE standard says for dynamic-sized types:
-     // runtime check
-     struct {bit a[$]; shortint b;} a = {{1,2,3,4}, 67};
-     int b = int'(a);
-     // compile time error
-     typedef struct {byte a[$]; bit b;} dest_t;
-     int a;
-     dest_t b = dest_t'(a);
-     */
-    auto width = type.getBitWidth();
-    if (width > 0)
-        return { 0, width };
-    if (type.isString())
-        return { 8, 0 };
-    bitwidth_t gcd = 0;
-    // TODO: bitwidth_t overflow
-    if (type.isUnpackedArray()) {
-        auto [gcd1, width1] = linearCoefficients(*type.getArrayElementType(), destination);
-        const auto& ct = type.getCanonicalType();
-        if (ct.kind == SymbolKind::FixedSizeUnpackedArrayType) {
-            auto rw = ct.as<FixedSizeUnpackedArrayType>().range.width();
-            gcd = gcd1 * rw;
-            width = width1 * rw;
-        }
-        else if (destination) {
-            if (destination > 1) {
-                if (!gcd1)
-                    gcd = width1;
-                else {
-                    gcd = gcd1;
-                    width = width1;
-                }
-            }
-        }
-        else
-            gcd = std::gcd(gcd1, width1);
-    }
-    else if (type.isUnpackedStruct()) {
-        auto& us = type.getCanonicalType().as<UnpackedStructType>();
-        for (auto& field : us.membersOfType<FieldSymbol>()) {
-            auto [gcd1, width1] = linearCoefficients(field.getType(), destination);
-            if (destination > 1 && gcd1 > 0)
-                destination--;
-            gcd = std::gcd(gcd, gcd1);
-            width += width1;
-        }
-    }
-    // TODO: classes
-    return { gcd, width };
-}
-
 bool Type::isCastCompatible(const Type& rhs, bool useBitstreamCast) const {
     // See [6.22.4] for Cast Compatible
     const Type* l = &getCanonicalType();
@@ -540,20 +487,24 @@ bool Type::isCastCompatible(const Type& rhs, bool useBitstreamCast) const {
     if (l->isAssignmentCompatible(*r))
         return true;
 
-    if (l->isEnum())
-        return r->isIntegral() || r->isFloating();
-
-    if (l->isString())
-        return r->isIntegral();
-
-    if (r->isString())
-        return l->isIntegral();
+    if (l->isEnum()) {
+        if (r->isIntegral() || r->isFloating())
+            return true;
+    }
+    else if (l->isString()) {
+        if (r->isIntegral())
+            return true;
+    }
+    else if (r->isString()) {
+        if (l->isIntegral())
+            return true;
+    }
 
     if (useBitstreamCast && l->isBitstreamType(true) && r->isBitstreamType()) {
         // bit-stream casting
+        ASSERT(l->isAggregate() || r->isAggregate());
         if (l->isFixedSize() && r->isFixedSize())
-            return l->bitstreamWidth() == r->bitstreamWidth() ||
-                   (!l->isAggregate() && !r->isAggregate());
+            return l->bitstreamWidth() == r->bitstreamWidth();
         else {
             auto [gcd0, fixed0] = linearCoefficients(*r);
             auto [gcd1, fixed1] = linearCoefficients(*l, 1);
@@ -568,124 +519,6 @@ bool Type::isCastCompatible(const Type& rhs, bool useBitstreamCast) const {
     }
 
     return false;
-}
-
-static ConstantValue constArray(const Type& type, span<const ConstantValue> elems) {
-    switch (type.kind) {
-        case SymbolKind::FixedSizeUnpackedArrayType:
-        case SymbolKind::DynamicArrayType:
-        case SymbolKind::UnpackedStructType:
-            return ConstantValue::Elements(elems.cbegin(), elems.cend());
-        case SymbolKind::QueueType:
-            return SVQueue(elems.cbegin(), elems.cend());
-        default:
-            return nullptr;
-    }
-}
-
-using PackVector = decltype(std::declval<ConstantValue>().bitstream());
-static SVInt slicePacked(PackVector::const_iterator& iter, bitwidth_t& bit, bitwidth_t width) {
-    const ConstantValue* cp;
-    for (;;) {
-        cp = *iter;
-        if (cp->isInteger()) {
-            if (bit < cp->integer().getBitWidth())
-                break;
-        }
-        else {
-            if (bit < cp->str().length() * 8)
-                break;
-        }
-        bit = 0;
-        iter++;
-    }
-    auto lsb = bit;
-    if (cp->isInteger()) {
-        const auto& ci = cp->integer();
-        auto msb = std::min(width, ci.getBitWidth());
-        bit += msb;
-        return ci.slice(static_cast<int32_t>(msb - 1), static_cast<int32_t>(lsb));
-    }
-    else {
-        const auto& str = cp->str();
-        auto byte0 = bit / 8;
-        auto byte1 = (bit + width - 1) / 8;
-        bitwidth_t len;
-        if (byte1 < str.length())
-            len = byte1 - byte0 + 1;
-        else {
-            len = static_cast<bitwidth_t>(str.length() - byte0);
-            width = len * 8 - bit % 8;
-        }
-        auto ci = SVInt(
-            width, span<const byte>(reinterpret_cast<const byte*>(str.data() + byte0), len), false);
-        bit += width;
-        if (lsb % 8 || (lsb + width) % 8)
-            return ci.slice(static_cast<int32_t>(width + 7 - (bit + width - 1) % 8), lsb % 8);
-        else
-            return ci;
-    }
-}
-
-static ConstantValue unpack(const Type& type, PackVector::const_iterator iter, bitwidth_t& bit,
-                            bitwidth_t& dynamic) {
-    if (type.isIntegral()) {
-        auto width = type.getBitWidth();
-        SmallVectorSized<SVInt, 8> buffer;
-        while (width > 0) {
-            auto ci = slicePacked(iter, bit, width);
-            width -= ci.getBitWidth();
-            if (!type.isFourState())
-                ci.flattenUnknowns();
-            buffer.emplace(ci);
-        }
-        auto cc = SVInt::concat(span<SVInt const>(buffer.begin(), buffer.end()));
-        cc.setSigned(type.isSigned());
-        return cc;
-    }
-    if (type.isString()) {
-        if (!dynamic)
-            return std::string();
-        auto width = dynamic;
-        ASSERT(width % 8 == 0);
-        dynamic = 0;
-        SmallVectorSized<SVInt, 8> buffer;
-        while (width > 0) {
-            auto ci = slicePacked(iter, bit, width);
-            width -= ci.getBitWidth();
-            ci.flattenUnknowns();
-            buffer.emplace(ci);
-        }
-        return ConstantValue(SVInt::concat(span<SVInt const>(buffer.begin(), buffer.end())))
-            .convertToStr();
-    }
-    if (type.isUnpackedArray()) {
-        const auto& ct = type.getCanonicalType();
-        SmallVectorSized<ConstantValue, 16> buffer;
-        if (ct.kind != SymbolKind::FixedSizeUnpackedArrayType) {
-            while (dynamic > 0)
-                buffer.emplace(unpack(*type.getArrayElementType(), iter, bit, dynamic));
-            ASSERT(!dynamic);
-        }
-        else {
-            const auto& ct1 = ct.as<FixedSizeUnpackedArrayType>();
-            const auto& elem = ct1.elementType;
-            for (auto width = ct1.elementType.bitstreamWidth() * ct1.range.width(); width > 0;
-                 width--)
-                buffer.emplace(unpack(elem, iter, bit, dynamic));
-        }
-        return constArray(ct, span<const ConstantValue>(buffer));
-    }
-    if (type.isUnpackedStruct()) {
-        SmallVectorSized<ConstantValue, 16> buffer;
-        const auto& ct = type.getCanonicalType();
-        auto& us = ct.as<UnpackedStructType>();
-        for (auto& field : us.membersOfType<FieldSymbol>())
-            buffer.emplace(unpack(field.getType(), iter, bit, dynamic));
-        return constArray(ct, span<const ConstantValue>(buffer));
-    }
-    // TODO: classes
-    return nullptr;
 }
 
 ConstantValue Type::bitstreamCast(const ConstantValue& value) const {
@@ -707,7 +540,13 @@ ConstantValue Type::bitstreamCast(const ConstantValue& value) const {
             dynmaic = srcSize - fixed2;
         }
     }
-    return unpack(*this, value.bitstream().cbegin(), bit, dynmaic);
+    const auto cv0 = value.bitstream();
+    auto iter = cv0.cbegin();
+    const auto cv = unpack(*this, iter, bit, dynmaic);
+    ASSERT(!dynmaic && bit == ((*iter)->isInteger() ? (*iter)->integer().getBitWidth()
+                                                    : (*iter)->str().length() * 8));
+    ASSERT(iter != cv0.cend() && ++iter == cv0.cend());
+    return cv;
 }
 
 bitmask<IntegralFlags> Type::getIntegralFlags() const {
